polyglotexchange/requirements.txt
Django==1.8.4
MySQL-python==1.2.5
South==1.0.2
argparse==1.2.1
decorator==4.0.2
django-markdown-deux==1.0.5
django-mongodb-engine==0.6.0
django-pagedown==0.1.0
django-tracking2==0.2.5
djangotoolbox==1.8.0
httplib2==0.9.1
ipython==4.0.0
ipython-genutils==0.1.0
markdown2==2.3.0
mongoengine==0.10.0
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
pymongo==3.0.3
pytz==2015.4
redis==2.10.3
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
twilio==4.5.0
wsgiref==0.1.2
